requirements.txt
flask
gunicorn
pillow
torch
transformers
accelerate
